Move RenderGraph from LigraControl to GraphItem.
