+++ Ligra.csproj
@@ -70,9 +70,6 @@
     <Compile Include="LigraControl.RenderExpression.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="LigraControl.RenderGraph.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
     <Compile Include="LigraControl.RenderMathPaint.cs">
       <SubType>UserControl</SubType>
     </Compile>
